finished db generic repository

+++ src/main/java/Repository/Db/DbRepository.java
@@ -1,10 +1,7 @@
 package Repository.Db;
-
 import Domain.Entity;
 import Domain.Filter;
 import Repository.IRepository;
-
-import java.security.PublicKey;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -12,15 +9,15 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-
 public abstract class DbRepository<ID,E extends Entity<ID>> implements IRepository<ID, E> {
-    Connection conn;
+    protected final Connection conn;
     public DbRepository(Connection conn) {
         this.conn = conn;
     }
     public Connection getConn() {
         return conn;
     }
+    @Override
     public List<E> filter(Filter filter){
         List<E> entities = new ArrayList<>();
         String where = filter.buildWhere();
@@ -53,19 +50,49 @@ public abstract class DbRepository<ID,E extends Entity<ID>> implements IReposito
         }
         return entities;
     }
-    public abstract String getTableName();
-    protected abstract E mapResultSetToEntity(ResultSet resultSet);
-    public Optional<E> remove(ID id){
+    @Override
+    public Optional<E> remove(ID id) {
+        Optional<E> entity = findById(id);
+        if (entity.isEmpty()) return Optional.empty();
         String sql = "DELETE FROM " + getTableName() + " WHERE id = ?";
-        try(var ps = conn.prepareStatement(sql)){
-            ps.setObject(1,id);
-            ResultSet resultSet = ps.executeQuery();
-            if(resultSet.next()){
-                return Optional.of(mapResultSetToEntity(resultSet));
+        try (PreparedStatement ps = conn.prepareStatement(sql)) {
+            ps.setObject(1, id);
+            ps.executeUpdate();
+            return entity;
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public Optional<E> findById(ID id) {
+        String sql = "SELECT * FROM " + getTableName() + " WHERE id = ?";
+        try (PreparedStatement ps = conn.prepareStatement(sql)) {
+            ps.setObject(1, id);
+            ResultSet rs = ps.executeQuery();
+            if (rs.next()) {
+                return Optional.of(mapResultSetToEntity(rs));
             }
             return Optional.empty();
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
     }
+    @Override
+    public Optional<E> update(E e) {
+        String sql = buildUpdateSql();
+        try (PreparedStatement ps = conn.prepareStatement(sql)) {
+            setUpdateParameters(ps, e);
+            int rows = ps.executeUpdate();
+            if (rows == 0)
+                return Optional.empty();
+            return Optional.of(e);
+        } catch (SQLException ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+    protected abstract String buildUpdateSql();
+    protected abstract void setUpdateParameters(PreparedStatement preparedStatement, E e) throws SQLException;
+    public abstract String getTableName();
+    protected abstract E mapResultSetToEntity(ResultSet resultSet);
 }
